Create README.md
B.Tech in Electronics and Computer Science Engineering | Aspiring Software Engineer.

Welcome to my GITHUB profile! I am currently pursuing my B.Tech from the School of Engineering at KIIT, specializing in Electronics and Computer Science Engineering. I am an ambitious and motivated individual seeking a valuable internship opportunity to enhance my skills, gain practical experience, and launch a successful career in the field of IT.

My passion for coding and software development has driven me to excel in my studies and explore various domains within the IT industry. I have a keen interest in software, web, and game development and aspire to work for a product-based company as a Software Engineer in the future. As a fresher, I am eager to immerse myself in the professional environment of an IT company and contribute my knowledge and skills to real-world projects.

As an individual, I possess a strong work ethic, attention to detail, and a disciplined approach to my responsibilities. I am always eager to take on new challenges and seize every opportunity for personal and professional growth. I believe in continuous learning and staying updated with the latest industry trends and technologies.

If you are interested in connecting with me regarding internship opportunities or any other position within your esteemed organization, please feel free to email me at shrutibose2004@gmail.com or connect with me on LinkedIn. I would be delighted to discuss how my skills and enthusiasm align with.

Thank you for visiting my profile, and I look forward to connecting with professionals who share a passion for technology and innovation.